Extract participant dashboard mapper into contest.participant.mapper.ts
The participant dashboard mapping (mapParticipantDashboardDto plus
9 internal helpers covering status, timeline, overview, paper/coding
report rows and details, trend points, and event feed) lived in
contest.mapper.ts but was self-contained. Moving it isolates ~210
lines into a focused module.

- New contest.participant.mapper.ts owns mapParticipantDashboardDto
  and its 9 private helpers; imports mapContestParticipantDto from
  contest.mapper.
- contest.mapper.ts re-exports mapParticipantDashboardDto so existing
  callers (repository, index) keep working.
- contest.mapper.ts: 676 -> 450 lines.
